Major update; bye bye filters
Ok this is a pretty major update, and hopefully will be the last time we
mess with this stuff. In this update I have gotten rid of filters. I
wasn't a big fan of them and they really didn't fit the OOP way of doing
things. Instead I have gone with the "more code but less messy" approach
of doing specialized handlers for everything. Check the file changes to
see what I mean. A new blog post and updates to the cookbook will be
coming soon.