Add EditedHandler and update to Crystal 0.35.0
This commit has a lot in it. First of all I added the `EditedHandler` and an `Edited` annotation to go with it. These make it possible to perform an action when a message is edited (perfect for spam filtering). I also added an `on_edit` parameter to the `CommandHandler` and `Command` annotation, allowing the handler to check edited messages for a command and act on it. An `edit` field was also added to the `CommandHandler::Context` to allow checking whether the handler was triggered because of an edit or not.

Also added were some small changes to make things work right with Crystal 0.35.0. Specifically logging has been changed. The `log_level` parameter has been moved from the `Client` constructor in favor of using `Log.setup`. That was really the only change needed for 0.35.0.